plugins.maya: Collect all types of atoms node
